requirements.txt
Flask==3.1.0
gunicorn==20.1.0
flask-cors==3.0.10
werkzeug==3.1.3
google-cloud-storage==3.1.0
tensorflow>=2.8.0
numpy
Pillow
opencv-python-headless>=4.5.0